[Bug #23] Fix TRAP handling code for nanomips user-mode
This was broken since 0bbc0396809f6caaaf96863dafe738e94f9b73ea
which disables the ISA mode-switch bit for nanoMIPS.

Though the ISA mode-switch bit is not required for address calculation
on nanoMIPS, the same MIPS_HFLAG_M16 was instrumental in decoding and
interpreting the TRAP instruction as well.  nanoMIPS has the same
instruction byte order as microMIPS, so instructions cannot be read in
simple byte-order, as is the case with MIPS16/MIPS32.

Distinct logic to extract trap-code from instruction is also required for nanoMIPS.

linux-user/
	* mips/cpu_loop.c (cpu_loop) <EXCP_TRAP>: Decode TRAP
	instruction for nanoMIPS by directly checking for the ISA.

Change-Id: I3dc985d6164454ef099a9bdfd4a131c78d6da039